requirements.txt
fastapi
uvicorn
google-generativeai
gtts
python-dotenv
playsound